郭长江-砍价完成-订单下单完成

+++ retail-modules/retail-bargain/retail-bargain-common/src/main/java/com.retail.bargain/domain/BargainRecordEntity.java
@@ -41,7 +41,7 @@ public class BargainRecordEntity implements Serializable {
 	/**
 	 * 砍价百分比
 	 */
-	private Double bargainOrderAccomplish;
+	private BigDecimal bargainOrderAccomplish;
 	/**
 	 * 砍价价格
 	 */
